added initial python support
Python support has been added to the configure.ac file to generate the
correct configure script which generates the headers and makefiles to
include python support. Python Runtime is currently in place of the JS
runtime, JS support must be enabled for the Python support to be
enabled as well.